+++ tests/ndarray/test_numpy.py
@@ -22,6 +22,6 @@ import blosc2
 def test_numpy(shape, chunks, blocks, dtype):
     size = int(np.prod(shape))
     nparray = np.arange(size, dtype=dtype).reshape(shape)
-    a = blosc2.asarray(nparray, chunks=chunks, blocks=blocks, dtype=dtype)
+    a = blosc2.asarray(nparray, chunks=chunks, blocks=blocks)
     nparray2 = a[...]
     np.testing.assert_almost_equal(nparray, nparray2)
